Set up gnucash.releases.xml to override the gnucash default
This file is meant to track package updates that are not related
to the gnucash release tarball, like updates to the flatpak
packaging scripts or updates to dependencies (like aqbanking)
between gnucash releases.

Package releases are typically named
<gnc.rel>-<xyz>, like 3.8-2
xyz is then the number of commits in the flathub repo since the first
gnucash <gnc.rel> release.